+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added `comp_vactube_sensor`, which allows detecting the presence of vactube objects.
 * Allow configuring various shadow/fast reflection options for vactube object ents.
 * Boolean keyvalues (yes/no) can now be set to `!$var`, to invert the value of the variable. 
+* Propcombine will now preserve prop fade distances, by calculating a new distance which encloses the original fade spheres.
 
 
 --------------------